requirements.txt
pytest
google-cloud-storage
pillow